+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/),
 and [PEP 440](https://www.python.org/dev/peps/pep-0440/).
 
+
+## [0.2.0] - 2023-XX-XX (not releaset yet)
+- new Pile constructor `openpile.construct.Pile.create_tubular` creating a 
+  circular and hollow steel pile of constant cross section.
+- new `openpile.construct.Pile` properties: `weight`, `volume`, 
+
 ## [0.1.0] - 2023-04-10
 ### Added
 - PISA sand and clay models (called Dunkirk_sand and Cowden_clay models)